requirements.txt
astroid==2.2.5
certifi==2019.9.11
chardet==3.0.4
Django==2.2.4
django-cors-headers==3.1.1
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pillow==6.2.0
pkg-resources==0.0.0
pycrypto==2.6.1
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
